requirements.txt
#backend
fastapi
uvicorn
ultralytics
opencv-python
python-multipart

# Frontend
streamlit
requests
Pillow

# Optional (if you want CORS support between frontend/backend)
python-multipart